Move chain/historicalState to chain/archiveStore/historicalState
