+++ packages/beacon-node/src/chain/archiveStore/historicalState/index.ts
